fix(goodix): GS32 typedef 還原 vendor 原樣 + goodix_mem_init 用 int32_t
之前為了讓 GCC 過,把 5 個 header / .c 的 GS32 從 'int' 改成 'long int',
結果反而打破 Keil。原因:int32_t 在 GCC arm-none-eabi 是 long int,在
armclang/Keil 是 int -- 兩個編譯器的 stdint 實作不同。

我把 GS32 改成 long int 後:
- GCC: int32_t = long int, GS32 = long int -- OK
- Keil: int32_t = int, GS32 = long int -- conflicting types

正解:GS32 維持 vendor 原樣 (int),只在實際衝突的 declaration 點
(gh30x_algo_hook.h 的 goodix_mem_init) 改用 int32_t 跟 goodix_mem.h
一致:
- GCC: int32_t = long int (both sides), GS32 = int 不參與這條宣告 -- OK
- Keil: int32_t = int (both sides), GS32 = int 不參與這條宣告 -- OK

兩個編譯器都過。

修改範圍:
- gh30x_algo_hook.h:328: extern GS32 goodix_mem_init(void*, GS32 size) →
  extern int32_t goodix_mem_init(void*, int32_t size)
- gh3018/gh30x_example_common.h, gh30x_hbd_ctrl.c, gh30x_algo/Basic/inc/
  {goodix_type.h, iot_sys_def.h}, gh30x_algo_demo/call/inc/goodix_algo.h
  全部 GS32/GU32 還原成 int/unsigned int (vendor 原樣)

+++ customer/peripherals/sensor/gh30x_algo_demo/call/inc/goodix_algo.h
@@ -14,8 +14,8 @@ typedef unsigned char       GU8;    /**< 8bit unsigned integer type */
 typedef signed char         GS8;    /**< 8bit signed integer type */
 typedef unsigned short      GU16;   /**< 16bit unsigned integer type */
 typedef signed short        GS16;   /**< 16bit signed integer type */
-typedef long int            GS32;   /**< 32bit signed integer type */
-typedef unsigned long int   GU32;   /**< 32bit unsigned integer type */
+typedef int                 GS32;   /**< 32bit signed integer type */
+typedef unsigned int        GU32;   /**< 32bit unsigned integer type */
 typedef float               GF32;   /**< float type */
 typedef double              GD64;   /**< double type */
 typedef char                GCHAR;  /**< char type */
